Update README.md to include installation instructions
I spent way too much time just trying to figure out how to install this. I know that it is more of a general tutorial that could apply to any Blender Add-on, however I got here from a tutorial that still assumed that this was a default blender Add-on (it should be).